==> 00_scripts_user_filter_24hs_delete/00_data_preprocessing/01_preprocessing_script_gather_all_dataframes_2019.ipynb <==
#!pip install pandas



# Gather all data

import csv
import json
import pandas as pd
from datetime import datetime
import numpy as np
from collections import Counter
from os.path import exists


import warnings
warnings.filterwarnings("ignore")


outputFolders = "../data_2019_processed/"

# Dataframe of all datasets
# initialize list of lists
datasets = [['0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs'],
            ['0tdmm6rwW3KquQ73ATYYJ5JkpMtvbppJ0VzA2GExdA'], 
            ['2cyV53lVyUtlMj0BRwilEWtYJwUiviYoL48cZBPBq0'], 
            ['2J22RukYnEbKTk7t+iUVDBkorcyL5NKN6TrLe89ys'], 
            #['5FLZBTVAPwdq9QezHE2sVCJIs7p+r6mCemA2gp9jATk'], #does not have the file
            ['7EYF5I04EVqisUJCVNHlqn77UAuOmwL2Dahxd3cA'], 
            ['a9Qgj8ENWrHvl9QqlXcIPKmyGMKgbfHk9Dbqon1HQP4'], 
            ['ae4JJBZDycEcY8McJF+3BxyvZ1619y03BNdCxzpZTc'], 
            ['Ch3u5Oaz96VSrQbf0z31X6jEIbeIekkC0mwPzCdeJ1U'], 
            ['CH8f0yZkZL13zWuE9ks1CkVJRVrr+jsGdUXHrZ6YeA'], 
            ['DHO1K4jgiwZJOfQTrxvKE2vn7hkjamigroGD5IaeRc'], 
            ['DHPqzSqSttiba1L3BD1cptNJPjSxZ8rXxF9mY3za6WA'], 
            ['dQEFscjqnIlug8Tgq97JohhSQPG2DEOWJqS86wCrcY'], 
            ['HFvs2CohmhHte+AaCzFasjzegGzxZKPhkrX23iI6Xo'], 
            ['jgB9E8v3Z6PKdTRTCMAijBllA9YEMtrmHbe4qsbmJWw'], 
            ['JkY++R7E8myldLN3on6iQ78Ee78zCbrLuggfwGju3I'], 
            ['K4SLohf+TN1Ak8Dn8iE3Lme7rEMPISfppB2sXfHX8'], 
            ['oGaWetJJJEWHuvYdWYo826SQxfhCExVVQ2da8LE1Y7Q'], 
            ['pyt24oiDAHsmgWMvkFKz2fn2pwcHiXchd6KchLM'], 
            ['PZCf1nfvhR+6fk+7+sPNMYOgb8BAMmtQtfoRS83Suc'], 
            ['QUNCATForxzK0HHw46LrGOMWh0eVA8Y5XWEiUXX+cQ'], 
            ['rIl2UK9+bQ+tzpFdbJAdbBxEa5GbgrgC030yEaENLw'], 
            ['RoBW3cDOO9wWRMPO2twQff83MPc+OXn6gJ+a1DafreI'], 
            ['SH3kQeyd5volraxw8vOyhlowNqWBPr1IJ9URNXUL4'], 
            ['VVpwFNMrEglveh6MDN8lrRzTy5OwzglD4FURfM4A2is'], 
            ['Wa1mcNmbh66S7VS6GIzyfCFMD3SGhbtDQyFP1ywJEsw'], 
            ['XCKRE0BWRHxfP1kZIihgtT+jUjSp2GE8v5ZlhcIhVmA'], 
            ['YI5Y79K6GXqAUoGP6PNyII8WKlAoel4urDxWSVVOvBw'], 
            ['ypklj+8GJ15rOIH1lpKQtFJOuK+VdvyCuBPqhY3aoM'], 
            ['ZSsAZ0Pq+MCqFrnjsRFn5Ua09pMCVaOV9c8ZuYb7XQY']]
  
# Create the pandas DataFrame
folders = pd.DataFrame(datasets, columns=['folder'])

#folders

pd.read_csv("../../Dataset_ECUADOR_2019/0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs/sleep_form_0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs.csv")

pd.read_csv("../data_2019_processed/form/sleep_form_0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs_treated.csv")

X = pd.read_csv("../data_2019_processed/form/sleep_form_0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs_treated.csv")

pd.to_datetime(X['timestamp'])

# Data pre-processing

allData = pd.DataFrame()
for indexFolder, folder in folders.iterrows():

    # check f the path exists (becouse kaggle removed the + character from file names)
    fileName = folder['folder']
    if(not(exists("../../Dataset_ECUADOR_2019/"+folder['folder']+"/"+fileName+".csv"))):
        fileName = folder['folder'].replace("+","")
    
    # data = pd.read_csv("../../Dataset_ECUADOR_2019/0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs/0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs.csv", delimiter=',', encoding="utf8")
    data1 = pd.read_csv("../../Dataset_ECUADOR_2019/"+folder['folder']+"/"+fileName+".csv",
                       delimiter=',', encoding="utf8")
    #data.to_csv("teste1.csv", sep=',', encoding='utf-8', index=False)
    data2 = pd.read_csv("../../Dataset_ECUADOR_2019/"+folder['folder']+"/Sleep_Detection_"+fileName+".csv",
                       delimiter=',', encoding="utf8")
    #dataFLabel = pd.read_csv("../Dataset_ECUADOR_2019/"+folder['folder']+"/sleep_form_"+folder['folder']+"_treated.csv",
    #               delimiter=',', encoding="utf8")
    dataFLabel = pd.read_csv("../data_2019_processed/form/sleep_form_"+folder['folder']+"_treated.csv",
                   delimiter=',', encoding="utf8")
    # merge data
    mergedDS = pd.merge(data1, data2, on='timestamp')
    print(folder['folder']+" "+str(mergedDS.shape[0]))
    # remove unecessary attributes not relevant to sleep detection: foregroundApp,wifi,bluetooth,ble, risk,suburb,municipality,city,town,county
    del mergedDS["foregroundApp"]
    del mergedDS["wifi"]
    del mergedDS["bluetooth"]
    del mergedDS["ble"]
    del mergedDS["risk"]
    del mergedDS["suburb"]
    del mergedDS["municipality"]
    del mergedDS["city"]
    del mergedDS["town"]
    del mergedDS["county"]
    del mergedDS["Unnamed: 0"]
    del mergedDS["activity_y"]
    # rename
    mergedDS.rename(columns = {'activity_x':'activity'}, inplace = True)
    #del mergedDS["Unnamed: 0_x"]
    #del mergedDS["Unnamed: 0_y"]
    # give the status of awake to all dataset
    mergedDS['class'] = 'awake'
    # convert label DF to date time
    dataFLabel['timestamp'] = pd.to_datetime(dataFLabel['timestamp'])
    mergedDS['timestamp2'] = 0
    for indexFolder, content in mergedDS['timestamp'].items():
        temp = datetime.strptime(content,'%Y-%m-%dT%H:%M:%SZ')
        mergedDS['timestamp2'][indexFolder] = temp.timestamp()
        
    #dataFLabel['sleep_time'] = dataFLabel['sleep_time'].astype('datetime64[ns]')
    #dataFLabel['wakeup_time'] = dataFLabel['wakeup_time'].astype('datetime64[ns]')
    # convert merged DF timestamp to datetime
    mergedDS['timestamp'] = pd.to_datetime(mergedDS['timestamp'])
    # time of the day
    #mergedDS['minutes_day'] = mergedDS['timestamp'].hour * 60 + mergedDS['timestamp'].minute
    mergedDS['minutes_day'] = mergedDS.timestamp.dt.hour * 60 + mergedDS.timestamp.dt.minute
    # initialize times
    mergedDS['sleep_time'] = 0
    mergedDS['wakeup_time'] = 0
    mergedDS['time_sleeping'] = 0
    # alter class to sleep based on sleep and wakeuptimes
    for index, labelSet in dataFLabel.iterrows():
        sleep_time = dataFLabel['sleep_time'][index]
        wakeup_time = dataFLabel['wakeup_time'][index]
        # replace value
        for indexT, contentT in mergedDS['timestamp2'].items():
            if(sleep_time <= contentT) and (contentT <= wakeup_time):
                mergedDS['class'][indexT] = 'asleep'
            if (contentT < wakeup_time and mergedDS['time_sleeping'][indexT] == 0) or dataFLabel.index[dataFLabel.shape[0]-1]==indexT:
                sleep_time_dt = datetime.fromtimestamp(sleep_time, tz = None)     
                wakeup_time_dt = datetime.fromtimestamp(wakeup_time, tz = None)        
                mergedDS['sleep_time'][indexT] = sleep_time_dt.hour * 60 + sleep_time_dt.minute 
                mergedDS['wakeup_time'][indexT] = wakeup_time_dt.hour * 60 + wakeup_time_dt.minute 
                mergedDS['time_sleeping'][indexT] = wakeup_time - sleep_time 
        #mergedDS['class'] = np.where((sleep_time <= mergedDS['timestamp2'])
        #                              and (mergedDS['timestamp2'] <= wakeup_time), 'asleep', mergedDS['class'])
        #mergedDS['sleep_time'] = sleep_time.hour * 60 + sleep_time.minute # in minutes
        #mergedDS['wakeup_time'] = wakeup_time.hour * 60 + wakeup_time.minute  # in minutes
        #mergedDS['time_sleeping'] = mergedDS['time_sleeping'] /60
    # remove temp information
    #del mergedDS['sleep_time']
    #del mergedDS['wakeup_time']
    #del mergedDS['time_sleeping']
    mergedDS['timestamp_text'] = mergedDS['timestamp']
    mergedDS['timestamp'] = mergedDS['timestamp2']
    del mergedDS['timestamp2']
    # replace the place from the class
    temp = mergedDS['class'].copy()
    del mergedDS['class']
    mergedDS['class'] = temp
    # print the guy
    mergedDS.to_csv(outputFolders+"student_"+folder['folder']+".csv", sep=',', encoding='utf-8', index=False)
    #mergedDS.to_csv("dados2019/student_"+folder['folder']+".csv", sep=',', encoding='utf-8', index=False)
    print(Counter(mergedDS['class']))
    # merge the data set if has both classes
    if('asleep' in Counter(mergedDS['class'])):
        if(allData.empty):
            allData = mergedDS
        else:
            #allData = allData.append(mergedDS, ignore_index=True)
            allData = pd.concat([allData, mergedDS])
        
# print all
#allData.to_csv(outputFolders+"allData.csv", sep=',', encoding='utf-8', index=False)
#allData.to_csv("dados2019/allData.csv", sep=',', encoding='utf-8', index=False)

allData

mergedDS
