==> 00_scripts_user_filter_24hs_delete/00_data_preprocessing/02_data_analysis_and_timestamp_filter.ipynb <==
import csv
import json
import pandas as pd
from datetime import datetime
import numpy as np
from os.path import exists

from collections import Counter

outputFolders = "../data_2019_processed/"


import warnings
warnings.filterwarnings("ignore")


#fileRaw = "allData.csv"
#fileNumeric = "-numeric.csv"
#fileNumericNormalized = "-numeric-normalized.csv"
#fileNumericClassification = "-numeric-classification.csv"
#fileNumericClassificationNormalized = "-classification-numeric-normalized.csv"

limitAwake = 64800 # 18 hs == 1080 min == 64800 s 
limitAwake = 43200 # 12 hs == 1080 min == 64800 s 

#limitAwake = 86400 # 24 hs == 1080 min == 64800 s 

# Dataframe of all datasets
# initialize list of lists
datasets = [['0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs'],
            ['0tdmm6rwW3KquQ73ATYYJ5JkpMtvbppJ0VzA2GExdA'], 
            ['2cyV53lVyUtlMj0BRwilEWtYJwUiviYoL48cZBPBq0'], 
            ['2J22RukYnEbKTk7t+iUVDBkorcyL5NKN6TrLe89ys'], 
            #['5FLZBTVAPwdq9QezHE2sVCJIs7p+r6mCemA2gp9jATk'], #does not have the file
            ['7EYF5I04EVqisUJCVNHlqn77UAuOmwL2Dahxd3cA'], 
            ['a9Qgj8ENWrHvl9QqlXcIPKmyGMKgbfHk9Dbqon1HQP4'], 
            ['ae4JJBZDycEcY8McJF+3BxyvZ1619y03BNdCxzpZTc'], 
            ['Ch3u5Oaz96VSrQbf0z31X6jEIbeIekkC0mwPzCdeJ1U'], 
            ['CH8f0yZkZL13zWuE9ks1CkVJRVrr+jsGdUXHrZ6YeA'], 
            ['DHO1K4jgiwZJOfQTrxvKE2vn7hkjamigroGD5IaeRc'], 
            ['DHPqzSqSttiba1L3BD1cptNJPjSxZ8rXxF9mY3za6WA'], 
            ['dQEFscjqnIlug8Tgq97JohhSQPG2DEOWJqS86wCrcY'], 
            ['HFvs2CohmhHte+AaCzFasjzegGzxZKPhkrX23iI6Xo'], 
            ['jgB9E8v3Z6PKdTRTCMAijBllA9YEMtrmHbe4qsbmJWw'], 
            ['JkY++R7E8myldLN3on6iQ78Ee78zCbrLuggfwGju3I'], 
            ['K4SLohf+TN1Ak8Dn8iE3Lme7rEMPISfppB2sXfHX8'], 
            ['oGaWetJJJEWHuvYdWYo826SQxfhCExVVQ2da8LE1Y7Q'], 
            ['pyt24oiDAHsmgWMvkFKz2fn2pwcHiXchd6KchLM'], 
            ['PZCf1nfvhR+6fk+7+sPNMYOgb8BAMmtQtfoRS83Suc'], 
            ['QUNCATForxzK0HHw46LrGOMWh0eVA8Y5XWEiUXX+cQ'], 
            ['rIl2UK9+bQ+tzpFdbJAdbBxEa5GbgrgC030yEaENLw'], 
            ['RoBW3cDOO9wWRMPO2twQff83MPc+OXn6gJ+a1DafreI'], 
            ['SH3kQeyd5volraxw8vOyhlowNqWBPr1IJ9URNXUL4'], 
            ['VVpwFNMrEglveh6MDN8lrRzTy5OwzglD4FURfM4A2is'], 
            ['Wa1mcNmbh66S7VS6GIzyfCFMD3SGhbtDQyFP1ywJEsw'], 
            ['XCKRE0BWRHxfP1kZIihgtT+jUjSp2GE8v5ZlhcIhVmA'], 
            ['YI5Y79K6GXqAUoGP6PNyII8WKlAoel4urDxWSVVOvBw'], 
            ['ypklj+8GJ15rOIH1lpKQtFJOuK+VdvyCuBPqhY3aoM'], 
            ['ZSsAZ0Pq+MCqFrnjsRFn5Ua09pMCVaOV9c8ZuYb7XQY']]
  
# Create the pandas DataFrame
folders = pd.DataFrame(datasets, columns=['folder'])

dataset = pd.read_csv(outputFolders+"student_oGaWetJJJEWHuvYdWYo826SQxfhCExVVQ2da8LE1Y7Q.csv")
#dataset = pd.read_csv(outputFolders+"student_ypklj+8GJ15rOIH1lpKQtFJOuK+VdvyCuBPqhY3aoM.csv")

dataset

#dataset.loc[dataset['class'] == "asleep"]
ff = dataset.loc[dataset['class'] == "asleep"]
ff.iloc[[2]] # use the order from the array
ff.loc[[380]] # use the index from the dataframe

#dataset.loc[dataset['class'] == "asleep"]
ff = dataset.loc[dataset['class'] == "asleep"]
#ff.iloc[[2]] # use the order from the array
ff
ff.iloc[0]['timestamp']

dataset['outside_range']  = 0
#dataset['nextAsleepTimeStamp']  = 0
#dataset['lastAsleepTimeStamp']  = 0
# first asleep timestamp
nextAsleepTimeStamp = ff.iloc[0]['timestamp']
lastAsleepTimeStamp = 0
count = 0
for index, content in dataset.iterrows():
    count = count + 1
    currentTimeStamp = content['timestamp']
    #dataset['nextAsleepTimeStamp'][index] = nextAsleepTimeStamp
    #dataset['lastAsleepTimeStamp'][index] = lastAsleepTimeStamp
    # outside_range before a timestamp
    if content['class'] == 'awake':
        if(currentTimeStamp < (nextAsleepTimeStamp-limitAwake) and currentTimeStamp > (lastAsleepTimeStamp+limitAwake) ):
            dataset['outside_range'][index] = 1
    else:
        # change of asleep timestamp
        if(index < dataset.shape[0]-1):
            if(dataset['class'][index+1] == 'awake'):
                lastAsleepTimeStamp = currentTimeStamp # updates the last
                # intermidiate variable
                tess = dataset.iloc[index+1:].loc[dataset['class'] == "asleep"]
                if(tess.shape[0] == 0): # used to treat the last nextAsleepTimeStamp
                    nextAsleepTimeStamp = 3790003633 # 2090-02-05 18:47:13
                else:
                    nextAsleepTimeStamp = dataset.iloc[index+1:].loc[dataset['class'] == "asleep"].iloc[0]['timestamp'] # updates the next asleep
                #print(nextAsleepTimeStamp)
                #print(dataset.iloc[index+1:].loc[dataset['class'] == "asleep"])
                #lastAsleepTimeStamp = currentTimeStamp # updates the last
                #nextAsleepTimeStamp = dataset.iloc[index+1:].loc[dataset['class'] == "asleep"].iloc[0]['timestamp'] # updates the next asleep

print(dataset.loc[dataset['outside_range'] == 1].shape[0])
print(dataset.loc[dataset['outside_range'] == 0].shape[0])
print(dataset.loc[dataset['outside_range'] == 1].shape[0])
print(dataset.loc[dataset['outside_range'] == 0].shape[0]-dataset.loc[dataset['class'] == "asleep"].shape[0])
print(dataset.loc[dataset['class'] == "asleep"].shape[0])
print(dataset.shape[0])
dataset.loc[dataset['outside_range'] == 1]



#dataset.loc[dataset['class'] == "asleep"]
dataset.iloc[1000:].loc[dataset['class'] == "asleep"].iloc[0]['timestamp']

#dataset.to_csv("test_lab.csv", sep=';', encoding='utf-8', index=False)

print("")
print("")
print("")
print("User: ","oGaWetJJJEWHuvYdWYo826SQxfhCExVVQ2da8LE1Y7Q")
print("")
print('location',Counter(dataset['location']))
print('activity',Counter(dataset['activity']))
print('day_of_week',Counter(dataset['day_of_week']))
print('light',Counter(dataset['light']),dataset['light'].sum())
print('sound',dataset['sound'].min(),dataset['sound'].sum())
print('time_sleeping',Counter(dataset['time_sleeping']))
print('sleep_time',Counter(dataset['sleep_time']))
print('wakeup_time',Counter(dataset['wakeup_time']))
print('proximity',Counter(dataset['proximity']))
print('class',Counter(dataset['class']))
print('sleep_time',Counter(dataset['sleep_time']))
print('wakeup_time',Counter(dataset['wakeup_time']))
print('timestamp',dataset['timestamp'][0],dataset['timestamp'][dataset.shape[0]-1])
print('time_to_next_alarm',Counter(dataset['time_to_next_alarm']))

print()

print("")
print("")
print("")

test = []

for indexFolder, folder in folders.iterrows():
    dataset = pd.read_csv(outputFolders+"student_"+folder['folder']+".csv")
    
    
    data = dict()
    # basic transformations 
    print("")
    print("")
    print("")
    print("student_"+str(folder['folder'])+".csv", Counter(dataset['class']) )
    print("")
    print('location',Counter(dataset['location']))
    print('activity',Counter(dataset['activity']))
    print('day_of_week',Counter(dataset['day_of_week']))
    print('light',dataset['light'].min(),dataset['light'].sum())
    print('sound',dataset['sound'].min(),dataset['sound'].sum())
    print('time_sleeping',Counter(dataset['time_sleeping']))
    print('sleep_time',Counter(dataset['sleep_time']))
    print('wakeup_time',Counter(dataset['wakeup_time']))
    print('proximity',Counter(dataset['proximity']))
    print('phone_lock',Counter(dataset['phone_lock']))
    print('class',Counter(dataset['class']))
    print('sleep_time',Counter(dataset['sleep_time']))
    print('wakeup_time',Counter(dataset['wakeup_time']))
    #print('timestamp',dataset['timestamp'][0],dataset['timestamp'][dataset.shape[0]-1])
    print("")
    #print('time_to_next_alarm',Counter(dataset['time_to_next_alarm']))
    print('time_to_next_alarm',dataset['time_to_next_alarm'].max())
    print("")
    print("")
    print("")
    
    
    #limitAwake = 64800 # 18 hs == 1080 min == 64800 s 
    #limitAwake = 86400
    #limitAfterAwake = 64800 # 18 hs == 1080 min == 64800 s 
    #limitBeforeSleep = 28800 # 8hs = 480 min = 28800 s\
    
    # ouside_range todos os dados com menos de timestamp - 64800
    dataset['outside_range']  = 0
    
    # first asleep timestamp
    nextAsleepTimeStamp = ff.iloc[0]['timestamp']
    lastAsleepTimeStamp = 0
    
    for index, content in dataset.iterrows():
        
        currentTimeStamp = content['timestamp']
        # outside_range before a timestamp
        if content['class'] == 'awake':
            if(currentTimeStamp < (nextAsleepTimeStamp-limitAwake) and currentTimeStamp > (lastAsleepTimeStamp+limitAwake) ):
                dataset['outside_range'][index] = 1
        else:
            # change of asleep timestamp
            if(index < dataset.shape[0]-1):
                if(dataset['class'][index+1] == 'awake'):
                    lastAsleepTimeStamp = currentTimeStamp # updates the last
                    # intermidiate variable
                    tess = dataset.iloc[index+1:].loc[dataset['class'] == "asleep"]
                    if(tess.shape[0] == 0): # used to treat the last nextAsleepTimeStamp
                        nextAsleepTimeStamp = 3790003633 # 2090-02-05 18:47:13
                    else:
                        # updates the next asleep
                        nextAsleepTimeStamp = dataset.iloc[index+1:].loc[dataset['class'] == "asleep"].iloc[0]['timestamp'] 
    
    #dataset = dataset.loc[dataset['outside_range'] == 0]
                       
    print("Total excluídos: ",dataset.loc[dataset['outside_range'] == 1].shape[0])
    print("Total incluídos: ",dataset.loc[dataset['outside_range'] == 0].shape[0])
    print("Total excluídos do awake: ",dataset.loc[dataset['outside_range'] == 0].shape[0]-dataset.loc[dataset['class'] == "asleep"].shape[0])
    print("Total asleep:",dataset.loc[dataset['class'] == "asleep"].shape[0])
    print("Total: ",dataset.shape[0])
    #dataset.loc[dataset['outside_range'] == 1]

    
    data["student"] = folder['folder']
    #data["timestamp_start"] = dataset['timestamp'][0]
    #data["timestamp_end"] = dataset['timestamp'][dataset.shape[0]-1]
    data["light_sum"] = dataset['light'].sum()
    data["sound_sum"] = dataset['sound'].sum()
    data["phone_lock"] = Counter(dataset['phone_lock'])
    data["location"] = Counter(dataset['location'])
    data["activity"] = Counter(dataset['activity'])
    data["day_of_week"] = Counter(dataset['day_of_week'])
    data["proximity"] = Counter(dataset['proximity'])
    data["time_to_next_alarm "] = dataset['time_to_next_alarm'].max()
    data["class "] = Counter(dataset['class'])
    
    test.append(data)

dataset.loc[dataset['outside_range'] == 0]

dataset

print("Starting script test")
print()
unprocessedAwake, unprocessedAsleep = 0,0
processedAwake, processedAsleep = 0,0
for indexFolder, folder in folders.iterrows():
    dataset = pd.read_csv(outputFolders+"student_"+folder['folder']+".csv")
    
    #limitAwake = 64800 # 18 hs == 1080 min == 64800 s 
    #limitAwake = 86400
    #limitAfterAwake = 64800 # 18 hs == 1080 min == 64800 s 
    #limitBeforeSleep = 28800 # 8hs = 480 min = 28800 s\
    
    # ouside_range todos os dados com menos de timestamp - 64800
    dataset['outside_range']  = 0
    
    # first asleep timestamp
    nextAsleepTimeStamp = ff.iloc[0]['timestamp']
    lastAsleepTimeStamp = 0
    
    for index, content in dataset.iterrows():
        
        currentTimeStamp = content['timestamp']
        # outside_range before a timestamp
        if content['class'] == 'awake':
            if(currentTimeStamp < (nextAsleepTimeStamp-limitAwake) and currentTimeStamp > (lastAsleepTimeStamp+limitAwake) ):
                dataset['outside_range'][index] = 1
        else:
            # change of asleep timestamp
            if(index < dataset.shape[0]-1):
                if(dataset['class'][index+1] == 'awake'):
                    lastAsleepTimeStamp = currentTimeStamp # updates the last
                    # intermidiate variable
                    tess = dataset.iloc[index+1:].loc[dataset['class'] == "asleep"]
                    if(tess.shape[0] == 0): # used to treat the last nextAsleepTimeStamp
                        nextAsleepTimeStamp = 3790003633 # 2090-02-05 18:47:13
                    else:
                        # updates the next asleep
                        nextAsleepTimeStamp = dataset.iloc[index+1:].loc[dataset['class'] == "asleep"].iloc[0]['timestamp'] 
    
    dataset2 = dataset.loc[dataset['outside_range'] == 0]
    
    data = dict()
    # basic transformations 
    #print("")
    #print("")
    print("")
    print("student_"+str(folder['folder'])+".csv", Counter(dataset['class']) )
    print("student_"+str(folder['folder'])+".csv", Counter(dataset2['class']) )
    
    unprocessedAwake = Counter(dataset['class'])['awake'] + unprocessedAwake
    processedAwake = Counter(dataset2['class'])['awake'] + processedAwake
    if("asleep" in Counter(dataset['class'])):
        unprocessedAsleep = Counter(dataset['class'])['asleep'] + unprocessedAsleep
        processedAsleep = Counter(dataset2['class'])['asleep'] + processedAsleep
print("Before pre-processing")
print("awake:",unprocessedAwake)
print("asleep:",unprocessedAsleep)
print("Total:",unprocessedAsleep+unprocessedAwake)
print()
print("After pre-processing")
print("awake:",processedAwake)
print("asleep:",processedAsleep)
print("Total:",processedAsleep+processedAwake)

unprocessed = unprocessedAsleep+unprocessedAwake
processed = processedAsleep+processedAwake
percent = (100/unprocessed*processed)
print("Total= 100 /",unprocessed,"*",processed)
print("Total=",percent,"%")
print("Total=",100-percent,"%")
print("Diference=",unprocessed-processed)
print()
print("from", 467889/77561,"times to",368651/77561,"times")
print()
print()
print("awake states from 3 datasets without sleep records: ",1398+59+1712)
print()
instancesToRemove = 1398+59+1712
unprocessed = unprocessedAsleep+unprocessedAwake-instancesToRemove
processed = processedAsleep+processedAwake-instancesToRemove
percent = (100/unprocessed*processed)
print("Total= 100 /",unprocessed,"*",processed)
print("Total=",percent,"%")
print("Total=",100-percent,"%")
print("Diference=",unprocessed-processed)
print()
print("from", (467889-instancesToRemove)/77561,"times to",(368651-instancesToRemove)/77561,"times")

unprocessed = unprocessedAsleep+unprocessedAwake
processed = processedAsleep+processedAwake
percent = (100/unprocessed*processed)
print("Total= 100 /",unprocessed,"*",processed)
print("Total=",percent,"%")
print("Total=",100-percent,"%")
print("Diference=",unprocessed-processed)
print()
print("from", 467889/77561,"times to",368651/77561,"times")
print()
print()
print("awake states from 3 datasets without sleep records: ",1398+59+1712)
print()
instancesToRemove = 1398+59+1712
unprocessed = unprocessedAsleep+unprocessedAwake-instancesToRemove
processed = processedAsleep+processedAwake-instancesToRemove
percent = (100/unprocessed*processed)
print("Total= 100 /",unprocessed,"*",processed)
print("Total=",percent,"%")
print("Total=",100-percent,"%")
print("Diference=",unprocessed-processed)
print()
print("from", (467889-instancesToRemove)/77561,"times to",(368651-instancesToRemove)/77561,"times")

unprocessedAwake, unprocessedAsleep = 0,0
processedAwake, processedAsleep = 0,0
for indexFolder, folder in folders.iterrows():
    dataset = pd.read_csv(outputFolders+"student_"+folder['folder']+".csv")
    
    #limitAwake = 64800 # 18 hs == 1080 min == 64800 s 
    #limitAwake = 86400 # 24hs
    #limitAfterAwake = 64800 # 18 hs == 1080 min == 64800 s 
    #limitBeforeSleep = 28800 # 8hs = 480 min = 28800 s\
    
    # ouside_range todos os dados com menos de timestamp - 64800
    dataset['outside_range']  = 0
    
    # first asleep timestamp
    nextAsleepTimeStamp = ff.iloc[0]['timestamp']
    lastAsleepTimeStamp = 0
    
    for index, content in dataset.iterrows():
        
        currentTimeStamp = content['timestamp']
        # outside_range before a timestamp
        if content['class'] == 'awake':
            if(currentTimeStamp < (nextAsleepTimeStamp-limitAwake) and currentTimeStamp > (lastAsleepTimeStamp+limitAwake) ):
                dataset['outside_range'][index] = 1
        else:
            # change of asleep timestamp
            if(index < dataset.shape[0]-1):
                if(dataset['class'][index+1] == 'awake'):
                    lastAsleepTimeStamp = currentTimeStamp # updates the last
                    # intermidiate variable
                    tess = dataset.iloc[index+1:].loc[dataset['class'] == "asleep"]
                    if(tess.shape[0] == 0): # used to treat the last nextAsleepTimeStamp
                        nextAsleepTimeStamp = 3790003633 # 2090-02-05 18:47:13
                    else:
                        # updates the next asleep
                        nextAsleepTimeStamp = dataset.iloc[index+1:].loc[dataset['class'] == "asleep"].iloc[0]['timestamp'] 
    
    dataset2 = dataset.loc[dataset['outside_range'] == 0]
    
    data = dict()
    # basic transformations 
    #print("")
    #print("")
    print("")
    print("student_"+str(folder['folder'])+".csv", Counter(dataset['class']) )
    print("student_"+str(folder['folder'])+"_filtred.csv", Counter(dataset2['class']) )
    
    unprocessedAwake = Counter(dataset['class'])['awake'] + unprocessedAwake
    processedAwake = Counter(dataset2['class'])['awake'] + processedAwake
    if("asleep" in Counter(dataset['class'])):
        unprocessedAsleep = Counter(dataset['class'])['asleep'] + unprocessedAsleep
        processedAsleep = Counter(dataset2['class'])['asleep'] + processedAsleep
        
        
    # delete data not useful
    del dataset2['sleep_time']
    del dataset2['wakeup_time']
    del dataset2['time_sleeping']
        
    dataset2.to_csv(outputFolders+"student_"+folder['folder']+"_filtred.csv", sep=',', encoding='utf-8', index=False)
print("Before pre-processing")
print("awake:",unprocessedAwake)
print("asleep:",unprocessedAsleep)
print("Total:",unprocessedAsleep+unprocessedAwake)
print()
print("After pre-processing")
print("awake:",processedAwake)
print("asleep:",processedAsleep)
print("Total:",processedAsleep+processedAwake)
