==> 00_scripts_user_filter_24hs_delete/00_data_preprocessing/06_DataCorrelation_oversample-SMOTE.ipynb <==
import csv
import json
import pandas as pd
from datetime import datetime
import numpy as np

mainfolder = "../data_2019_processed/"
file_name_modifier = "_smote"
outputFolders = ""

# take the list of directories and concat them
def loadDataFromFolders(foldersToLoad,inputFolders,fileType = ""):
    print(len(foldersToLoad), "datasets")
    for i in range(0,len(foldersToLoad)):
        currentFolder = foldersToLoad[i]
        print(i , "-", currentFolder,inputFolders+"student_"+currentFolder+"_transformed"+fileType+".csv")
        #print(trainingDataSet[i])
        if(i == 0):
            temp_data = pd.read_csv(inputFolders+"student_"+currentFolder+"_transformed"+fileType+".csv")
        else:
            dataset = pd.read_csv(inputFolders+"student_"+currentFolder+"_transformed"+fileType+".csv")
            temp_data = pd.concat([temp_data, dataset])
    # return the dataset        
    return temp_data

# client datasets used on the training process (75% of data)
trainFolders =  ['0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs',
                '0tdmm6rwW3KquQ73ATYYJ5JkpMtvbppJ0VzA2GExdA', 
                '2cyV53lVyUtlMj0BRwilEWtYJwUiviYoL48cZBPBq0', 
                '2J22RukYnEbKTk7t+iUVDBkorcyL5NKN6TrLe89ys', 
                #['5FLZBTVAPwdq9QezHE2sVCJIs7p+r6mCemA2gp9jATk'], #does not have the file
                '7EYF5I04EVqisUJCVNHlqn77UAuOmwL2Dahxd3cA', 
                'a9Qgj8ENWrHvl9QqlXcIPKmyGMKgbfHk9Dbqon1HQP4', 
                'ae4JJBZDycEcY8McJF+3BxyvZ1619y03BNdCxzpZTc', 
                'Ch3u5Oaz96VSrQbf0z31X6jEIbeIekkC0mwPzCdeJ1U', 
                'CH8f0yZkZL13zWuE9ks1CkVJRVrr+jsGdUXHrZ6YeA', 
                'DHO1K4jgiwZJOfQTrxvKE2vn7hkjamigroGD5IaeRc', 
                #'DHPqzSqSttiba1L3BD1cptNJPjSxZ8rXxF9mY3za6WA', # does not have asleep data
                'dQEFscjqnIlug8Tgq97JohhSQPG2DEOWJqS86wCrcY', 
                'HFvs2CohmhHte+AaCzFasjzegGzxZKPhkrX23iI6Xo', 
                'jgB9E8v3Z6PKdTRTCMAijBllA9YEMtrmHbe4qsbmJWw', 
                'JkY++R7E8myldLN3on6iQ78Ee78zCbrLuggfwGju3I', 
                'K4SLohf+TN1Ak8Dn8iE3Lme7rEMPISfppB2sXfHX8', 
                'oGaWetJJJEWHuvYdWYo826SQxfhCExVVQ2da8LE1Y7Q', 
                'pyt24oiDAHsmgWMvkFKz2fn2pwcHiXchd6KchLM', 
                #'PZCf1nfvhR+6fk+7+sPNMYOgb8BAMmtQtfoRS83Suc', # does not have asleep data
                'QUNCATForxzK0HHw46LrGOMWh0eVA8Y5XWEiUXX+cQ', 
                #'rIl2UK9+bQ+tzpFdbJAdbBxEa5GbgrgC030yEaENLw', 
                #'RoBW3cDOO9wWRMPO2twQff83MPc+OXn6gJ+a1DafreI', 
                'SH3kQeyd5volraxw8vOyhlowNqWBPr1IJ9URNXUL4']
                #'VVpwFNMrEglveh6MDN8lrRzTy5OwzglD4FURfM4A2is', 
                #'Wa1mcNmbh66S7VS6GIzyfCFMD3SGhbtDQyFP1ywJEsw', 
                #'XCKRE0BWRHxfP1kZIihgtT+jUjSp2GE8v5ZlhcIhVmA', 
                #'YI5Y79K6GXqAUoGP6PNyII8WKlAoel4urDxWSVVOvBw', 
                #'ypklj+8GJ15rOIH1lpKQtFJOuK+VdvyCuBPqhY3aoM', 
                #'ZSsAZ0Pq+MCqFrnjsRFn5Ua09pMCVaOV9c8ZuYb7XQY']
            
# client datasets used on the training process (25% of data)
testFolders =  [#'0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs',
                #'0tdmm6rwW3KquQ73ATYYJ5JkpMtvbppJ0VzA2GExdA', 
                #'2cyV53lVyUtlMj0BRwilEWtYJwUiviYoL48cZBPBq0', 
                #'2J22RukYnEbKTk7t+iUVDBkorcyL5NKN6TrLe89ys', 
                #['5FLZBTVAPwdq9QezHE2sVCJIs7p+r6mCemA2gp9jATk'], #does not have the file
                #'7EYF5I04EVqisUJCVNHlqn77UAuOmwL2Dahxd3cA', 
                #'a9Qgj8ENWrHvl9QqlXcIPKmyGMKgbfHk9Dbqon1HQP4', 
                #'ae4JJBZDycEcY8McJF+3BxyvZ1619y03BNdCxzpZTc', 
                #'Ch3u5Oaz96VSrQbf0z31X6jEIbeIekkC0mwPzCdeJ1U', 
                #'CH8f0yZkZL13zWuE9ks1CkVJRVrr+jsGdUXHrZ6YeA', 
                #'DHO1K4jgiwZJOfQTrxvKE2vn7hkjamigroGD5IaeRc', 
                #'DHPqzSqSttiba1L3BD1cptNJPjSxZ8rXxF9mY3za6WA', # does not have asleep data
                #'dQEFscjqnIlug8Tgq97JohhSQPG2DEOWJqS86wCrcY', 
                #'HFvs2CohmhHte+AaCzFasjzegGzxZKPhkrX23iI6Xo', 
                #'jgB9E8v3Z6PKdTRTCMAijBllA9YEMtrmHbe4qsbmJWw', 
                #'JkY++R7E8myldLN3on6iQ78Ee78zCbrLuggfwGju3I', 
                #'K4SLohf+TN1Ak8Dn8iE3Lme7rEMPISfppB2sXfHX8', 
                #'oGaWetJJJEWHuvYdWYo826SQxfhCExVVQ2da8LE1Y7Q', 
                #'pyt24oiDAHsmgWMvkFKz2fn2pwcHiXchd6KchLM', 
                #'PZCf1nfvhR+6fk+7+sPNMYOgb8BAMmtQtfoRS83Suc', # does not have asleep data
                #'QUNCATForxzK0HHw46LrGOMWh0eVA8Y5XWEiUXX+cQ', 
                'rIl2UK9+bQ+tzpFdbJAdbBxEa5GbgrgC030yEaENLw', 
                'RoBW3cDOO9wWRMPO2twQff83MPc+OXn6gJ+a1DafreI', 
                #'SH3kQeyd5volraxw8vOyhlowNqWBPr1IJ9URNXUL4'] 
                'VVpwFNMrEglveh6MDN8lrRzTy5OwzglD4FURfM4A2is', 
                'Wa1mcNmbh66S7VS6GIzyfCFMD3SGhbtDQyFP1ywJEsw', 
                'XCKRE0BWRHxfP1kZIihgtT+jUjSp2GE8v5ZlhcIhVmA', 
                'YI5Y79K6GXqAUoGP6PNyII8WKlAoel4urDxWSVVOvBw', 
                'ypklj+8GJ15rOIH1lpKQtFJOuK+VdvyCuBPqhY3aoM', 
                'ZSsAZ0Pq+MCqFrnjsRFn5Ua09pMCVaOV9c8ZuYb7XQY']

# client datasets used on the training process (75% of data)
allFolders =  ['0Jf4TH9Zzse0Z1Jjh7SnTOe2MMzeSnFi7feTnkG6vgs',
                '0tdmm6rwW3KquQ73ATYYJ5JkpMtvbppJ0VzA2GExdA', 
                '2cyV53lVyUtlMj0BRwilEWtYJwUiviYoL48cZBPBq0', 
                '2J22RukYnEbKTk7t+iUVDBkorcyL5NKN6TrLe89ys', 
                #['5FLZBTVAPwdq9QezHE2sVCJIs7p+r6mCemA2gp9jATk'], #does not have the file
                '7EYF5I04EVqisUJCVNHlqn77UAuOmwL2Dahxd3cA', 
                'a9Qgj8ENWrHvl9QqlXcIPKmyGMKgbfHk9Dbqon1HQP4', 
                'ae4JJBZDycEcY8McJF+3BxyvZ1619y03BNdCxzpZTc', 
                'Ch3u5Oaz96VSrQbf0z31X6jEIbeIekkC0mwPzCdeJ1U', 
                'CH8f0yZkZL13zWuE9ks1CkVJRVrr+jsGdUXHrZ6YeA', 
                'DHO1K4jgiwZJOfQTrxvKE2vn7hkjamigroGD5IaeRc', 
                #'DHPqzSqSttiba1L3BD1cptNJPjSxZ8rXxF9mY3za6WA', # does not have asleep data
                'dQEFscjqnIlug8Tgq97JohhSQPG2DEOWJqS86wCrcY', 
                'HFvs2CohmhHte+AaCzFasjzegGzxZKPhkrX23iI6Xo', 
                'jgB9E8v3Z6PKdTRTCMAijBllA9YEMtrmHbe4qsbmJWw', 
                'JkY++R7E8myldLN3on6iQ78Ee78zCbrLuggfwGju3I', 
                'K4SLohf+TN1Ak8Dn8iE3Lme7rEMPISfppB2sXfHX8', 
                'oGaWetJJJEWHuvYdWYo826SQxfhCExVVQ2da8LE1Y7Q', 
                'pyt24oiDAHsmgWMvkFKz2fn2pwcHiXchd6KchLM', 
                #'PZCf1nfvhR+6fk+7+sPNMYOgb8BAMmtQtfoRS83Suc', # does not have asleep data
                'QUNCATForxzK0HHw46LrGOMWh0eVA8Y5XWEiUXX+cQ', 
                'rIl2UK9+bQ+tzpFdbJAdbBxEa5GbgrgC030yEaENLw', 
                'RoBW3cDOO9wWRMPO2twQff83MPc+OXn6gJ+a1DafreI', 
                'SH3kQeyd5volraxw8vOyhlowNqWBPr1IJ9URNXUL4',
                'VVpwFNMrEglveh6MDN8lrRzTy5OwzglD4FURfM4A2is', 
                'Wa1mcNmbh66S7VS6GIzyfCFMD3SGhbtDQyFP1ywJEsw', 
                'XCKRE0BWRHxfP1kZIihgtT+jUjSp2GE8v5ZlhcIhVmA', 
                'YI5Y79K6GXqAUoGP6PNyII8WKlAoel4urDxWSVVOvBw', 
                'ypklj+8GJ15rOIH1lpKQtFJOuK+VdvyCuBPqhY3aoM', 
                'ZSsAZ0Pq+MCqFrnjsRFn5Ua09pMCVaOV9c8ZuYb7XQY']

# All data correlation

dataframe = loadDataFromFolders(allFolders,mainfolder,file_name_modifier)
#dataframe = pd.read_csv("dados2019/student_a9Qgj8ENWrHvl9QqlXcIPKmyGMKgbfHk9Dbqon1HQP4.csv", delimiter=',', encoding="utf8")

# Exibe info
print(dataframe.info())

# convert data and show the new info
# activity
dataframe['activity'] = dataframe['activity'].replace(['Unknown'], '1')
dataframe['activity'] = dataframe['activity'].replace(['Still'], '2')
dataframe['activity'] = dataframe['activity'].replace(['Tilting'], '3')
dataframe['activity'] = dataframe['activity'].replace(['Exercise'], '4')
dataframe['activity'] = dataframe['activity'].replace(['In vehicle'], '5')
# class
dataframe['class'] = dataframe['class'].replace(['awake'], '0')
dataframe['class'] = dataframe['class'].replace(['asleep'], '1')
# day_of_week
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Tuesday'], '1')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Wednesday'], '2')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Thursday'], '3')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Friday'], '4')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Saturday'], '5')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Sunday'], '6')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Monday'], '7')
# phone_lock
dataframe['phone_lock'] = dataframe['phone_lock'].replace(['not locked'], '0')
dataframe['phone_lock'] = dataframe['phone_lock'].replace(['locked'], '1')
# location
dataframe['location'] = dataframe['location'].replace(['Other'], '1')
dataframe['location'] = dataframe['location'].replace(['University'], '2')
dataframe['location'] = dataframe['location'].replace(['House'], '3')

# timestamp
if 'timestamp2' in dataframe.columns:
    dataframe['timestamp_text'] = dataframe['timestamp']
    dataframe['timestamp'] = dataframe['timestamp2']
    del dataframe['timestamp2']
#dataframe['timestamp'] = datetime.strptime(dataframe['timestamp'],'%Y-%m-%d %H:%M:%S').timestamp() * 1000
#dataframe['timestamp'] = pd.to_datetime(dataframe['timestamp'],'%Y-%m-%d %H:%M:%S') * 1000
#dataframe['timestamp'] = dataframe['timestamp'].dt.second()
# activity
dataframe['activity'] = dataframe['activity'].astype(np.integer) 
dataframe['class'] = dataframe['class'].astype(np.integer) 
dataframe['day_of_week'] = dataframe['day_of_week'].astype(np.integer) 
dataframe['phone_lock'] = dataframe['phone_lock'].astype(np.integer) 
dataframe['location'] = dataframe['location'].astype(np.integer) 
# Exibe info
print(dataframe.info())# convert values

dataframe

 #'pearson', "   "'spearman', 'kendall'
a = dataframe['class'].corr(dataframe['activity'])
b = dataframe['class'].corr(dataframe['location'])
#c = dataframe['class'].corr(dataframe['distanceHome'])
#d = dataframe['class'].corr(dataframe['showActuation'])
e = dataframe['class'].corr(dataframe['timestamp'])
f = dataframe['class'].corr(dataframe['day_of_week'])
g = dataframe['class'].corr(dataframe['light'])
h = dataframe['class'].corr(dataframe['phone_lock'])
i = dataframe['class'].corr(dataframe['proximity'])
j = dataframe['class'].corr(dataframe['sound'])
l = dataframe['class'].corr(dataframe['time_to_next_alarm'])
#m = dataframe['class'].corr(dataframe['sleep_time'])
#n = dataframe['class'].corr(dataframe['wakeup_time'])
p = dataframe['class'].corr(dataframe['minutes_day'])
o = dataframe['class'].corr(dataframe['class'])
print("Correlations")
corrrrrrr = (("activity",a),
            ("location",b),
      #      ("distanceHome",c),
     #       ("showActuation",d),
            ("timestamp",e),
            ("day_of_week",f),
            ("light",g),
            ("phone_lock",h),
            ("proximity",i),
            ("sound",j),
#            ("sleep_time",m),
#            ("wakeup_time",n),
            ("time_to_next_alarm",l),
            ("minutes_day",p),
            ("class",o))
print (corrrrrrr)

type(corrrrrrr)

tests = pd.DataFrame(corrrrrrr,columns=("Feature","Correlation"))
tests.sort_values(by=['Correlation'])

# Only training correlation

outputFolders = "train/"

dataframe = loadDataFromFolders(trainFolders,mainfolder,file_name_modifier)
#dataframe = pd.read_csv("dados2019/student_a9Qgj8ENWrHvl9QqlXcIPKmyGMKgbfHk9Dbqon1HQP4.csv", delimiter=',', encoding="utf8")

# convert data and show the new info
# activity
dataframe['activity'] = dataframe['activity'].replace(['Unknown'], '1')
dataframe['activity'] = dataframe['activity'].replace(['Still'], '2')
dataframe['activity'] = dataframe['activity'].replace(['Tilting'], '3')
dataframe['activity'] = dataframe['activity'].replace(['Exercise'], '4')
dataframe['activity'] = dataframe['activity'].replace(['In vehicle'], '5')
# class
dataframe['class'] = dataframe['class'].replace(['awake'], '0')
dataframe['class'] = dataframe['class'].replace(['asleep'], '1')
# day_of_week
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Tuesday'], '1')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Wednesday'], '2')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Thursday'], '3')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Friday'], '4')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Saturday'], '5')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Sunday'], '6')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Monday'], '7')
# phone_lock
dataframe['phone_lock'] = dataframe['phone_lock'].replace(['not locked'], '0')
dataframe['phone_lock'] = dataframe['phone_lock'].replace(['locked'], '1')
# location
dataframe['location'] = dataframe['location'].replace(['Other'], '1')
dataframe['location'] = dataframe['location'].replace(['University'], '2')
dataframe['location'] = dataframe['location'].replace(['House'], '3')

# timestamp
if 'timestamp2' in dataframe.columns:
    dataframe['timestamp_text'] = dataframe['timestamp']
    dataframe['timestamp'] = dataframe['timestamp2']
    del dataframe['timestamp2']
#dataframe['timestamp'] = datetime.strptime(dataframe['timestamp'],'%Y-%m-%d %H:%M:%S').timestamp() * 1000
#dataframe['timestamp'] = pd.to_datetime(dataframe['timestamp'],'%Y-%m-%d %H:%M:%S') * 1000
#dataframe['timestamp'] = dataframe['timestamp'].dt.second()
# activity
dataframe['activity'] = dataframe['activity'].astype(np.integer) 
dataframe['class'] = dataframe['class'].astype(np.integer) 
dataframe['day_of_week'] = dataframe['day_of_week'].astype(np.integer) 
dataframe['phone_lock'] = dataframe['phone_lock'].astype(np.integer) 
dataframe['location'] = dataframe['location'].astype(np.integer) 
# Exibe info
print(dataframe.info())# convert values

 #'pearson', "   "'spearman', 'kendall'
a = dataframe['class'].corr(dataframe['activity'])
b = dataframe['class'].corr(dataframe['location'])
#c = dataframe['class'].corr(dataframe['distanceHome'])
#d = dataframe['class'].corr(dataframe['showActuation'])
#e = dataframe['class'].corr(dataframe['timestamp'])
f = dataframe['class'].corr(dataframe['day_of_week'])
g = dataframe['class'].corr(dataframe['light'])
h = dataframe['class'].corr(dataframe['phone_lock'])
i = dataframe['class'].corr(dataframe['proximity'])
j = dataframe['class'].corr(dataframe['sound'])
l = dataframe['class'].corr(dataframe['time_to_next_alarm'])
#m = dataframe['class'].corr(dataframe['sleep_time'])
#n = dataframe['class'].corr(dataframe['wakeup_time'])
p = dataframe['class'].corr(dataframe['minutes_day'])
o = dataframe['class'].corr(dataframe['class'])
print("Correlations")
corrrrrrr = (("activity",a),
            ("location",b),
           # ("distanceHome",c),
           # ("showActuation",d),
           # ("timestamp",e),
            ("day_of_week",f),
            ("light",g),
            ("phone_lock",h),
            ("proximity",i),
            ("sound",j),
#            ("sleep_time",m),
#            ("wakeup_time",n),
            ("time_to_next_alarm",l),
            ("minutes_day",p),
            ("class",o))
print (corrrrrrr)

tests = pd.DataFrame(corrrrrrr,columns=("Feature","Correlation"))
tests.sort_values(by=['Correlation'])

# Test Folder

dataframe = loadDataFromFolders(testFolders,mainfolder,file_name_modifier)
#dataframe = pd.read_csv("dados2019/student_a9Qgj8ENWrHvl9QqlXcIPKmyGMKgbfHk9Dbqon1HQP4.csv", delimiter=',', encoding="utf8")

# convert data and show the new info
# activity
dataframe['activity'] = dataframe['activity'].replace(['Unknown'], '1')
dataframe['activity'] = dataframe['activity'].replace(['Still'], '2')
dataframe['activity'] = dataframe['activity'].replace(['Tilting'], '3')
dataframe['activity'] = dataframe['activity'].replace(['Exercise'], '4')
dataframe['activity'] = dataframe['activity'].replace(['In vehicle'], '5')
# class
dataframe['class'] = dataframe['class'].replace(['awake'], '0')
dataframe['class'] = dataframe['class'].replace(['asleep'], '1')
# day_of_week
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Tuesday'], '1')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Wednesday'], '2')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Thursday'], '3')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Friday'], '4')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Saturday'], '5')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Sunday'], '6')
dataframe['day_of_week'] = dataframe['day_of_week'].replace(['Monday'], '7')
# phone_lock
dataframe['phone_lock'] = dataframe['phone_lock'].replace(['not locked'], '0')
dataframe['phone_lock'] = dataframe['phone_lock'].replace(['locked'], '1')
# location
dataframe['location'] = dataframe['location'].replace(['Other'], '1')
dataframe['location'] = dataframe['location'].replace(['University'], '2')
dataframe['location'] = dataframe['location'].replace(['House'], '3')

# timestamp
if 'timestamp2' in dataframe.columns:
    dataframe['timestamp_text'] = dataframe['timestamp']
    dataframe['timestamp'] = dataframe['timestamp2']
    del dataframe['timestamp2']
#dataframe['timestamp'] = datetime.strptime(dataframe['timestamp'],'%Y-%m-%d %H:%M:%S').timestamp() * 1000
#dataframe['timestamp'] = pd.to_datetime(dataframe['timestamp'],'%Y-%m-%d %H:%M:%S') * 1000
#dataframe['timestamp'] = dataframe['timestamp'].dt.second()
# activity
dataframe['activity'] = dataframe['activity'].astype(np.integer) 
dataframe['class'] = dataframe['class'].astype(np.integer) 
dataframe['day_of_week'] = dataframe['day_of_week'].astype(np.integer) 
dataframe['phone_lock'] = dataframe['phone_lock'].astype(np.integer) 
dataframe['location'] = dataframe['location'].astype(np.integer) 
# Exibe info
print(dataframe.info())# convert values

#'pearson', "   "'spearman', 'kendall'
a = dataframe['class'].corr(dataframe['activity'])
b = dataframe['class'].corr(dataframe['location'])
#c = dataframe['class'].corr(dataframe['distanceHome'])
#d = dataframe['class'].corr(dataframe['showActuation'])
#e = dataframe['class'].corr(dataframe['timestamp'])
f = dataframe['class'].corr(dataframe['day_of_week'])
g = dataframe['class'].corr(dataframe['light'])
h = dataframe['class'].corr(dataframe['phone_lock'])
i = dataframe['class'].corr(dataframe['proximity'])
j = dataframe['class'].corr(dataframe['sound'])
l = dataframe['class'].corr(dataframe['time_to_next_alarm'])
#m = dataframe['class'].corr(dataframe['sleep_time'])
#n = dataframe['class'].corr(dataframe['wakeup_time'])
p = dataframe['class'].corr(dataframe['minutes_day'])
o = dataframe['class'].corr(dataframe['class'])
print("Correlations")
corrrrrrr = (("activity",a),
            ("location",b),
           # ("distanceHome",c),
           # ("showActuation",d),
           # ("timestamp",e),
            ("day_of_week",f),
            ("light",g),
            ("phone_lock",h),
            ("proximity",i),
            ("sound",j),
#            ("sleep_time",m),
#            ("wakeup_time",n),
            ("time_to_next_alarm",l),
            ("minutes_day",p),
            ("class",o))
print (corrrrrrr)

tests = pd.DataFrame(corrrrrrr,columns=("Feature","Correlation"))
tests.sort_values(by=['Correlation'])
